postgis-sqls/pyproject.toml
[project]
name = "postgis-sqls"
version = "0.1.0"
description = "한국 행정동 경계 데이터를 PostGIS 데이터베이스에 임포트하는 도구"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "geopandas>=0.13.2",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
    "shapely>=2.0.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "isort>=5.0.0",
    "flake8>=4.0.0",
]
analysis = [
    "pandas>=1.3.0",
    "geopandas>=0.10.0",
    "folium>=0.12.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 88
